requirements.txt
Flask~=3.1.0
paho-mqtt~=2.1.0
requests~=2.32.3
flask_sqlalchemy
flask_login
mysql-connector-python
bcrypt~=4.2.1
pyaml
gunicorn  # 可选，用于生产环境
pytest~=8.3.5
PyYAML~=6.0.2
Flask-Login~=0.6.3
Werkzeug~=3.1.3
Flask-JWT-Extended~=4.7.1
SQLAlchemy~=2.0.38
Flask-Migrate~=4.1.0
flask-cors~=5.0.1
dotenv~=0.9.9
python-dotenv~=1.1.0